A beautiful project with layered architecture and clean code
In our project, it is a good example of writing clean code that is based on list generics, libraries are used, and is suitable for its layered architecture.We can easily understand the formation of nested forms in this project.An update will come to the project and then it will be much more comprehensive, but for now, you can easily save and list operations. I am very happy to share the project with you.